requirements.txt
flask
gtts
qrcode
pdfplumber
pillow